refactor: HabitItem 패키지 분리
